+++ .gitignore
@@ -13,6 +13,7 @@
 /geoportal-application/geoportal-harvester-war/target/
 /geoportal-commons/target/
 /geoportal-commons/geoportal-commons-utils/target/
+/geoportal-commons/geoportal-commons-pdf/target/
 /geoportal-commons/geoportal-commons-robots/target/
 /geoportal-commons/geoportal-commons-csw-client/target/
 /geoportal-commons/geoportal-commons-gpt-client/target/
