+++ geoportal-commons/geoportal-commons-utils/pom.xml
@@ -37,15 +37,5 @@
             <artifactId>commons-text</artifactId>
             <version>1.2</version>
         </dependency>
-        <dependency>
-            <groupId>org.apache.pdfbox</groupId>
-            <artifactId>pdfbox</artifactId>
-            <version>2.0.9</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.pdfbox</groupId>
-            <artifactId>xmpbox</artifactId>
-            <version>2.0.9</version>
-        </dependency>
     </dependencies>
 </project>
