+++ geoportal-connectors/geoportal-harvester-unc/src/main/java/com/esri/geoportal/harvester/unc/UncFile.java
@@ -32,7 +32,7 @@ import com.esri.geoportal.commons.meta.MapAttribute;
 import com.esri.geoportal.commons.meta.MetaException;
 import com.esri.geoportal.commons.meta.util.WKAConstants;
 import com.esri.geoportal.commons.meta.xml.SimpleDcMetaBuilder;
-import com.esri.geoportal.commons.utils.PdfUtils;
+import com.esri.geoportal.commons.pdf.PdfUtils;
 import com.esri.geoportal.commons.utils.XmlUtils;
 import com.esri.geoportal.harvester.api.base.SimpleDataReference;
